sounds that do not carry through walls will not carry through floors / ceilings

+++ code/game/objects/items/stacks/stack.dm
@@ -18,6 +18,7 @@
 	var/singular_name
 	var/amount = 1
 	var/max_amount = 50 //also see stack recipes initialisation, param "max_res_amount" must be equal to this max_amount
+	var/is_cyborg = 0 // It's 1 if module is used by a cyborg, and uses its storage
 	var/datum/robot_energy_storage/source
 	var/cost = 1 // How much energy from storage it costs
 	var/merge_type = null // This path and its children should merge with this stack, defaults to src.type
@@ -34,6 +35,9 @@
 		grind_results[grind_results[i]] *= get_amount() //Gets the key at position i, then the reagent amount of that key, then multiplies it by stack size
 
 /obj/item/stack/grind_requirements()
+	if(is_cyborg)
+		to_chat(usr, span_warning("[src] is electronically synthesized in my chassis and can't be ground up!"))
+		return
 	return TRUE
 
 /obj/item/stack/Initialize(mapload, new_amount, merge = TRUE)
@@ -44,6 +48,9 @@
 		new type(loc, max_amount, FALSE)
 	if(!merge_type)
 		merge_type = type
+	if(custom_materials && custom_materials.len)
+		for(var/i in custom_materials)
+			custom_materials[getmaterialref(i)] = mats_per_stack * amount
 	. = ..()
 	if(merge)
 		for(var/obj/item/stack/S in loc)
@@ -51,6 +58,13 @@
 				merge(S)
 	var/list/temp_recipes = get_main_recipes()
 	recipes = temp_recipes.Copy()
+	if(material_type)
+		var/datum/material/M = getmaterialref(material_type) //First/main material
+		for(var/i in M.categories)
+			switch(i)
+				if(MAT_CATEGORY_RIGID)
+					var/list/temp = SSmaterials.rigid_stack_recipes.Copy()
+					recipes += temp
 	update_weight()
 	update_icon()
 
@@ -86,6 +100,12 @@
 
 /obj/item/stack/examine(mob/user)
 	. = ..()
+	if (is_cyborg)
+		if(singular_name)
+			. += "There is enough energy for [get_amount()] [singular_name]\s."
+		else
+			. += "There is enough energy for [get_amount()]."
+		return
 	if(singular_name)
 		if(get_amount()>1)
 			. += "There are [get_amount()] [singular_name]\s in the stack."
@@ -98,7 +118,10 @@
 	. += span_notice("Alt-click to take a custom amount.")
 
 /obj/item/stack/proc/get_amount()
-	. = (amount)
+	if(is_cyborg)
+		. = round(source.energy / cost)
+	else
+		. = (amount)
 
 /obj/item/stack/attack_self(mob/user)
 //	interact(user)
@@ -169,7 +192,7 @@
 	if (href_list["sublist"] && !href_list["make"])
 		interact(usr, text2num(href_list["sublist"]))
 	if (href_list["make"])
-		if (get_amount() < 1)
+		if (get_amount() < 1 && !is_cyborg)
 			qdel(src)
 
 		var/list/recipes_list = recipes
@@ -203,6 +226,25 @@
 			O.setDir(usr.dir)
 		use(R.req_amount * multiplier)
 
+		if(R.applies_mats && custom_materials && custom_materials.len)
+			var/list/used_materials = list()
+			for(var/i in custom_materials)
+				used_materials[getmaterialref(i)] = R.req_amount / R.res_amount * (MINERAL_MATERIAL_AMOUNT / custom_materials.len)
+			O.set_custom_materials(used_materials)
+
+		//START: oh fuck i'm so sorry
+		if(istype(O, /obj/structure/windoor_assembly))
+			var/obj/structure/windoor_assembly/W = O
+			W.ini_dir = W.dir
+		else if(istype(O, /obj/structure/window))
+			var/obj/structure/window/W = O
+			W.ini_dir = W.dir
+		//END: oh fuck i'm so sorry
+
+		else if(istype(O, /obj/item/restraints/handcuffs/cable))
+			var/obj/item/cuffs = O
+			cuffs.color = color
+
 		if (QDELETED(O))
 			return //It's a stack and has already been merged
 
@@ -219,26 +261,34 @@
 /obj/item/stack/proc/building_checks(datum/stack_recipe/R, multiplier)
 	if (get_amount() < R.req_amount*multiplier)
 		if (R.req_amount*multiplier>1)
-			to_chat(usr, "<span class='warning'>I haven't got enough [src] to build \the [R.req_amount*multiplier] [R.title]\s!</span>")
+			to_chat(usr, span_warning("I haven't got enough [src] to build \the [R.req_amount*multiplier] [R.title]\s!"))
 		else
-			to_chat(usr, "<span class='warning'>I haven't got enough [src] to build \the [R.title]!</span>")
+			to_chat(usr, span_warning("I haven't got enough [src] to build \the [R.title]!"))
 		return FALSE
 	var/turf/T = get_turf(usr)
 
+	var/obj/D = R.result_type
+	if(R.window_checks && !valid_window_location(T, initial(D.dir) == FULLTILE_WINDOW_DIR ? FULLTILE_WINDOW_DIR : usr.dir))
+		to_chat(usr, span_warning("The [R.title] won't fit here!"))
+		return FALSE
 	if(R.one_per_turf && (locate(R.result_type) in T))
-		to_chat(usr, "<span class='warning'>There is another [R.title] here!</span>")
+		to_chat(usr, span_warning("There is another [R.title] here!"))
 		return FALSE
 	if(R.on_floor)
 		if(!isfloorturf(T))
-			to_chat(usr, "<span class='warning'>\The [R.title] must be constructed on the floor!</span>")
+			to_chat(usr, span_warning("\The [R.title] must be constructed on the floor!"))
 			return FALSE
 		for(var/obj/AM in T)
 			if(istype(AM,/obj/structure/grille))
 				continue
 			if(istype(AM,/obj/structure/table))
 				continue
+			if(istype(AM,/obj/structure/window))
+				var/obj/structure/window/W = AM
+				if(!W.fulltile)
+					continue
 			if(AM.density)
-				to_chat(usr, "<span class='warning'>Theres a [AM.name] here. You cant make a [R.title] here!</span>")
+				to_chat(usr, span_warning("Theres a [AM.name] here. You cant make a [R.title] here!"))
 				return FALSE
 	if(R.placement_checks)
 		switch(R.placement_checks)
@@ -247,17 +297,19 @@
 				for(var/direction in GLOB.cardinals)
 					step = get_step(T, direction)
 					if(locate(R.result_type) in step)
-						to_chat(usr, "<span class='warning'>\The [R.title] must not be built directly adjacent to another!</span>")
+						to_chat(usr, span_warning("\The [R.title] must not be built directly adjacent to another!"))
 						return FALSE
 			if(STACK_CHECK_ADJACENT)
 				if(locate(R.result_type) in range(1, T))
-					to_chat(usr, "<span class='warning'>\The [R.title] must be constructed at least one tile away from others of its type!</span>")
+					to_chat(usr, span_warning("\The [R.title] must be constructed at least one tile away from others of its type!"))
 					return FALSE
 	return TRUE
 
 /obj/item/stack/use(used, transfer = FALSE, check = TRUE) // return 0 = borked; return 1 = had enough
 	if(check && zero_amount())
 		return FALSE
+	if (is_cyborg)
+		return source.use_charge(used * cost)
 	if (amount < used)
 		return FALSE
 	amount -= used
@@ -282,13 +334,22 @@
 	return TRUE
 
 /obj/item/stack/proc/zero_amount()
+	if(is_cyborg)
+		return source.energy < cost
 	if(amount < 1)
 		qdel(src)
 		return 1
 	return 0
 
 /obj/item/stack/proc/add(amount)
-	src.amount += amount
+	if (is_cyborg)
+		source.add_charge(amount * cost)
+	else
+		src.amount += amount
+	if(custom_materials && custom_materials.len)
+		for(var/i in custom_materials)
+			custom_materials[getmaterialref(i)] = MINERAL_MATERIAL_AMOUNT * src.amount
+		set_custom_materials() //Refresh
 	update_icon()
 	update_weight()
 
@@ -296,7 +357,10 @@
 	if(QDELETED(S) || QDELETED(src) || S == src) //amusingly this can cause a stack to consume itself, let's not allow that.
 		return
 	var/transfer = get_amount()
-	transfer = min(transfer, S.max_amount - S.amount)
+	if(S.is_cyborg)
+		transfer = min(transfer, round((S.source.max_energy - S.source.energy) / S.cost))
+	else
+		transfer = min(transfer, S.max_amount - S.amount)
 	if(pulledby)
 		pulledby.start_pulling(S)
 	S.copy_evidences(src)
@@ -309,7 +373,7 @@
 		merge(o)
 	. = ..()
 
-/obj/item/stack/hitby(atom/movable/AM, skipcatch, hitpush, blocked, datum/thrownthing/throwingdatum, damage_type = "blunt")
+/obj/item/stack/hitby(atom/movable/AM, skipcatch, hitpush, blocked, datum/thrownthing/throwingdatum, d_type = "blunt")
 	if(istype(AM, merge_type))
 		merge(AM)
 	. = ..()
@@ -328,6 +392,8 @@
 	if(isturf(loc)) // to prevent people that are alt clicking a tile to see its content from getting undesidered pop ups
 		return
 	if(!istype(user) || !user.canUseTopic(src, BE_CLOSE, ismonkey(user)))
+		return
+	if(is_cyborg)
 		return
 	else
 		if(zero_amount())
@@ -370,6 +436,10 @@
 	add_hiddenprint_list(from.return_hiddenprints())
 	fingerprintslast  = from.fingerprintslast
 	//TODO bloody overlay
+
+/obj/item/stack/microwave_act(obj/machinery/microwave/M)
+	if(istype(M) && M.dirty < 100)
+		M.dirty += amount
 
 /*
  * Recipe datum
